switch jest configs to typescript.  might as well be typescript everywhere
