--- jest.config.js
@@ -1,6 +1,0 @@
-const baseConfig = require("./jest.config.base");
-
-module.exports = {
-  ...baseConfig,
-  projects: ["<rootDir>/packages/*"],
-};
