+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "typecheck": "tsc",
     "format": "prettier -w ./*.js ./*.ts packages",
-    "lint": "eslint packages",
+    "lint": "eslint ./*.ts ./*.js packages",
     "publish-devtools": "cd $REPLAY_DIR/devtools && cd packages/shared && npx --yes yalc publish --private && cd ../../packages/protocol && npx yalc publish --private && cd ../../packages/replay-next && npx yalc publish --private",
     "link-devtools": "npx yalc link shared && npx yalc link protocol && npx yalc link replay-next",
     "yalc-devtools": "publish-devtools && link-devtools",
@@ -50,6 +50,7 @@
     "prettier": "^3.4.1",
     "prettier-plugin-import-sort": "^0.0.7",
     "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
     "tsconfig-paths-jest": "^0.0.1",
     "tsx": "^4.19.2",
