+++ packages/replay-data/src/recording-data/DependencyGraph.ts
@@ -3,12 +3,12 @@ import { ExecutionPoint } from "@replayio/protocol";
 import merge from "lodash/merge";
 import orderBy from "lodash/orderBy";
 
+import { DGAnalyzeDependencies } from "../backend-wrapper/analysis/dependency-graph";
 import {
   AnalyzeDependenciesResult,
   DependencyChainStep,
   DependencyGraphMode,
 } from "../backend-wrapper/analysis/dg-types";
-import { DGAnalyzeDependencies } from "../backend-wrapper/analysis/dependency-graph";
 import PointQueries from "./PointQueries";
 import ReplaySession from "./ReplaySession";
 import { CodeAtPoint, FrameWithPoint } from "./types";
