+++ packages/replay-data/src/recording-data/bindings/DynamicScopes.ts
@@ -1,6 +1,4 @@
 /**
- * 
+ *
  */
-export default class DynamicScopes {
-
-}
+export default class DynamicScopes {}
