+++ packages/source-parser/jest.config.ts
@@ -1,0 +1,9 @@
+import baseConfig from "../../jest.config.base";
+import type { Config } from "@jest/types";
+
+const config: Config.InitialOptions = {
+  ...baseConfig,
+  rootDir: __dirname
+};
+
+export default config;
